eks-agent/requirements.txt
boto3
mcp
mcp-proxy
pydantic-core
pygments
setuptools
strands-agents
strands-agents-tools
bedrock-agentcore
urllib3
uv
wheel
